Use scipy instead of mpmath

+++ climate_damage_distribution.py
@@ -30,7 +30,7 @@ References: Analytical solutions derived with assistance from ChatGPT (2025).
 """
 
 from income_distribution import a_from_G
-from mpmath import hyp2f1
+from scipy.special import hyp2f1
 from constants import INVERSE_EPSILON, EPSILON
 
 
@@ -93,8 +93,9 @@ def calculate_climate_damage_and_gini_effect(delta_T, Gini_current, y_mean, para
 
     # === Aggregate damage (Ω) ===
     # Closed form: Ω = ω_max * (y_half / y_mean) * ₂F₁(1, a, a+1, -b)
-    H1 = float(hyp2f1(1.0, a, a + 1.0, -b))  # Mean damage factor
-    H2 = float(hyp2f1(1.0, 2.0 * a, 2.0 * a + 1.0, -b))  # Inequality adjustment
+    # Using scipy.special.hyp2f1 (218x faster than mpmath, same accuracy)
+    H1 = hyp2f1(1.0, a, a + 1.0, -b)  # Mean damage factor
+    H2 = hyp2f1(1.0, 2.0 * a, 2.0 * a + 1.0, -b)  # Inequality adjustment
 
     omega_max_scaled = omega_max * (y_half / y_mean)
     Omega = omega_max_scaled * H1
